requirements_all_except_cbnet.txt
# pip install -r requirements_all_except_cbnet.txt

# base ----------------------------------------
# Cython
opencv-python
matplotlib
numpy
Pillow
PyYAML>=5.3.1
scipy
tensorboard>=1.5
tqdm>=4.41.0
joblib
jupyter
ipywidgets
openmim
ensemble-boxes
shapely

# logging -------------------------------------
# wandb

# plotting ------------------------------------
seaborn
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.8.1
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
pycocotools # COCO mAP
